+++ src/Cwaff-Guns/IronMaid.cs
@@ -55,8 +55,8 @@ public class IronMaid : CwaffGun
     {
         base.OnSwitchedToThisGun();
         LaunchAllBullets(this.PlayerOwner);
-        //NOTE: this might be a vanilla bug where automatically reloads while guns are switched out don't reload secondary modules
-        //TODO: patch ForceImmediateReload()
+        //NOTE: this might be a vanilla bug where automatic reloads while guns are switched out don't reload secondary modules
+        //TODO: can possibly patch ForceImmediateReload() to fix this
         if (this.Mastered && this.gun.ClipShotsRemaining == this.gun.DefaultModule.numberOfShotsInClip)
             foreach (ProjectileModule mod in this.gun.Volley.projectiles)
                 if (this.gun.m_moduleData.TryGetValue(mod, out ModuleShootData msd))