+++ changelog.md
@@ -2,8 +2,9 @@
 
 ## 1.22.3 (TBD)
 
-- Added mastery for Chroma
+- Added masteries for Chroma and Flakseed
 - Fixed issue with certain non-looping VFX animations preventing other unrelated animations from playing properly
+- Adjusted shader lighting on Flakseed's flowers
 
 ## 1.22.2 (2025-03-17)